feat: implemented declaration and assignment skeleton.

+++ src/symbol_table.rs
@@ -70,7 +70,7 @@ impl ScopeInfo<'_> {
 
     pub fn add(&mut self, ident: &Ident, type_: Type) -> MessageResult<Ident> {
         // increase the space needed by the stack frame by the size of the given type
-        self.symbol_table.size += type_.size();
+        self.symbol_table.size += type_.size() as u32;
 
         // manage new offset of the given variable
         let offset = self.symbol_table.size;
